dotrep-v2/services/anomaly_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymysql==1.1.1
cryptography==43.0.1
pandas==2.2.3
scikit-learn==1.6.0
numpy==2.1.3
python-dotenv==1.0.1
pydantic==2.10.3


